--- iOS/Core/SharedUserDefaultsKeys.swift
@@ -1,27 +1,0 @@
-//
-//  SharedUserDefaultsKeys.swift
-//  DuckDuckGo
-//
-//  Copyright © 2025 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-/// These keys are shared between the main app and extensions, like the widgets.
-public enum SharedUserDefaultsKeys {
-
-    public static let groupName = "group.com.duckduckgo.app"
-
-    public static let isAIChatEnabled = "aichat.settings.isEnabled"
-
-}
